requirements.txt
certifi
cryptography
dateutils
django-auxilium
django-braces
django-configurations
django-countries
django-crispy-forms
django-pipeline
django-vanilla-views
djangorestframework
gunicorn
markdown
newrelic
opbeat
psycopg2
six

https://github.com/miki725/breach_buster/archive/master.tar.gz
